refactor: 테스트 메소드 분리
