+++ src/test/java/wooteco/subway/admin/acceptance/LineAcceptanceTest.java
@@ -53,4 +53,24 @@ public class LineAcceptanceTest extends AcceptanceTest {
         List<LineResponse> linesAfterDelete = getLines();
         assertThat(linesAfterDelete.size()).isEqualTo(3);
     }
+
+    List<LineResponse> getLines() {
+        return get("/lines").jsonPath().getList(".", LineResponse.class);
+    }
+
+    void updateLine(Long id, LocalTime startTime, LocalTime endTime) {
+        Map<String, String> params = new HashMap<>();
+        params.put("startTime", startTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("endTime", endTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+        params.put("intervalTime", "10");
+
+        update(
+                "/lines/" + id,
+                params
+        );
+    }
+
+    void deleteLine(Long id) {
+        delete("/lines/" + id);
+    }
 }
